moving address utils to it's own file
