+++ src/utils.ts
@@ -3,11 +3,8 @@ import _qrcode from 'qrcode-generator';
 
 import Reactotron from 'reactotron-react-native';
 import {u8aConcat} from '@polkadot/util';
-import {ADDRESS_PREFIX_KUSAMA, ADDRESS_PREFIX_LITENTRY, ADDRESS_PREFIX_POLKADOT, FRAME_SIZE} from './constants';
-import {checkAddress, isEthereumChecksum} from '@polkadot/util-crypto';
+import {FRAME_SIZE} from './constants';
 import {FunctionMetadataLatest} from '@polkadot/types/interfaces';
-import {AccountAddressType, NetworkType} from './types';
-import {trim} from 'lodash';
 
 const MULTIPART = new Uint8Array([0]);
 
@@ -89,38 +86,8 @@ export function rawDataToU8A(rawData: string): Uint8Array | null {
   return bytes;
 }
 
-export function parseAddress(payload: string): AccountAddressType {
-  const parts = trim(payload).split(':').filter(Boolean);
-
-  if (parts.length === 1) {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    return {protocol: '', address: parts[0]!, name: ''};
-  }
-
-  if (parts.length !== 4) {
-    throw new Error('address format wrong');
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-  return {protocol: parts[0], address: parts[1]!, name: parts[3]!};
-}
-
 export const ReactotronDebug = Reactotron.debug;
 
-export const isAddressValid = (network: NetworkType, address: string) => {
-  switch (network.key) {
-    case 'polkadot':
-      return checkAddress(address, ADDRESS_PREFIX_POLKADOT)[0];
-    case 'kusama':
-      return checkAddress(address, ADDRESS_PREFIX_KUSAMA)[0];
-    case 'litentry_test':
-      return checkAddress(address, ADDRESS_PREFIX_LITENTRY)[0];
-    case 'ethereum':
-      return isEthereumChecksum(address); // fixme
-    default:
-      return false;
-  }
-};
 export const WHITESPACE = [' ', '\t'];
 
 export function validateFormField(
